+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/BaseRandomTest.java
@@ -294,6 +294,7 @@ public abstract class BaseRandomTest {
       throws IOException, ClassNotFoundException, SeedException {
     // Serialise an RNG.
     final BaseRandom rng = createRng();
+    rng.setSeedGenerator(getTestSeedGenerator());
     RandomTestUtils.assertEquivalentWhenSerializedAndDeserialized(rng);
   }